pyproject.toml
[tool.poetry]
name = "rollandbar"
version = "1.0.0"
description = ""
authors = ["n.korolkov <nick1994209@yandex.ru>"]

[tool.poetry.dependencies]
python = "^3.8"
django = "^3.0.5"
djangorestframework = "^3.11.0"
# for generating swagger and openapi schema
uritemplate = "^3.0.1"
PyYAML = "^5.3.1"
Pillow = "^7.1.1"

[tool.poetry.dev-dependencies]
flake8 = "^3.7.9"
flake8-commas = "^2.0.0"
flake8-broken-line = "^0.2.0"
flake8-docstrings = "^1.5.0"
flake8-isort = "^2.9.1"
flake8-print = "^3.1.4"
flake8-quotes = "^3.0.0"
flake8-string-format = "^0.3.0"
unify = "^0.5"
isort = "^4.3.21"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
